requirements.txt
yfinance>=0.2.40
feedparser>=6.0.11
praw>=7.7.1
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
